requirements.txt
sentence-transformers
faiss-cpu
tiktoken
numpy
requests
beautifulsoup4
tqdm